refactor(core): split coordinator by invariant owner
